fix: remove redundant f32::from() to satisfy clippy

+++ crates/terminal/src/terminal.rs
@@ -210,12 +210,12 @@ impl TerminalBounds {
         // Tolerance to prevent f32 precision from losing a row:
         // `N * line_height / line_height` can be N-epsilon, which floor()
         // would round down, pushing the first line into invisible scrollback.
-        let raw = f32::from(self.bounds.size.height / self.line_height);
+        let raw = self.bounds.size.height / self.line_height;
         raw.next_up().floor() as usize
     }
 
     pub fn num_columns(&self) -> usize {
-        let raw = f32::from(self.bounds.size.width / self.cell_width);
+        let raw = self.bounds.size.width / self.cell_width;
         raw.next_up().floor() as usize
     }
